+++ android/src/main/kotlin/com/pubmatic/sdk/flutter_openwrap_sdk/banner/POBBannerViewClient.kt
@@ -139,5 +139,9 @@ class POBBannerViewClient(
     override fun onAdClicked(bannerView: POBBannerView) {
       channel.invokeMethod("onAdClicked", POBUtils.getArgumentMap(adId))
     }
+
+    override fun onAdImpression(view: POBBannerView) {
+      channel.invokeMethod("onAdImpression", POBUtils.getArgumentMap(adId))
+    }
   }
 }