+++ android/src/main/kotlin/com/pubmatic/sdk/flutter_openwrap_sdk/rewarded/POBRewardedClient.kt
@@ -141,5 +141,9 @@ class POBRewardedClient(
       arguments["amount"] = reward.amount
       channel.invokeMethod("onReceiveReward", arguments)
     }
+
+    override fun onAdImpression(rewardedAd: POBRewardedAd) {
+      channel.invokeMethod("onAdImpression", POBUtils.getArgumentMap(adId))
+    }
   }
 }