+++ example/lib/screens/mrec_display/openwrap_mrec_display_screen.dart
@@ -37,7 +37,7 @@ class _OpenWrapMRECDisplayScreen extends State<OpenWrapMRECDisplayScreen> {
     OpenWrapSDK.setApplicationInfo(applicationInfo);
 
     // Initialise banner ad.
-    // For test IDs refer - https://community.pubmatic.com/display/IOPO/Test+and+debug+your+integration
+    // For test IDs refer - https://help.pubmatic.com/openwrap/docs/test-and-debug-your-integration-5#test-profileplacements
     _bannerAd = POBBannerAd(
         pubId: pubId,
         profileId: profileId,
@@ -111,4 +111,9 @@ class _BannerAdListener implements POBBannerAdListener {
   @override
   POBAdEvent<POBBannerAd>? get onAdClicked =>
       (POBBannerAd ad) => developer.log('$_tag: Ad Clicked');
+
+  /// Callback method Notifies that the banner impression occurred.
+  @override
+  POBAdEvent<POBBannerAd>? get onAdImpression =>
+      (POBBannerAd ad) => developer.log('$_tag: Ad Impression');
 }